+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/test_factories.ts
@@ -1,5 +1,5 @@
 import { data } from './data/resource.js';
-import { defaultNodeFunc, node16Func } from './function.js';
+import { defaultNodeFunc, funcWithSsm, node16Func } from './function.js';
 import { storage } from './storage/resource.js';
 import { auth } from './auth/resource.js';
 
@@ -9,4 +9,5 @@ export const dataStorageAuthWithTriggers = {
   defaultNodeFunc,
   data,
   node16Func,
+  funcWithSsm,
 };