+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aws-amplify/backend-auth
 
+## 1.0.2
+
+### Patch Changes
+
+- 1dc61de: bumping consumers of auth-construct
+- Updated dependencies [147fc86]
+  - @aws-amplify/auth-construct@1.1.1
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [15c2b7c]
+  - @aws-amplify/auth-construct@1.1.0
+  - @aws-amplify/backend-output-storage@1.0.1
+
 ## 1.0.0
 
 ### Major Changes