+++ packages/backend-output-schemas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/backend-output-schemas
 
+## 1.1.0
+
+### Minor Changes
+
+- 15c2b7c: remove client ids from auth output schema
+
 ## 1.0.0
 
 ### Major Changes