+++ packages/client-config/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/client-config
 
+## 1.0.2
+
+### Patch Changes
+
+- f7b8089: Correctly generate Gen1 social provider config form Gen2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [15c2b7c]
+  - @aws-amplify/backend-output-schemas@1.1.0
+  - @aws-amplify/deployed-backend-client@1.0.1
+  - @aws-amplify/model-generator@1.0.1
+
 ## 1.0.0
 
 ### Major Changes