+++ packages/model-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/model-generator
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [15c2b7c]
+  - @aws-amplify/backend-output-schemas@1.1.0
+  - @aws-amplify/deployed-backend-client@1.0.1
+
 ## 1.0.0
 
 ### Major Changes